+++ package.json
@@ -101,7 +101,7 @@
         }
       },
       {
-        "files": "encryptActions/*",
+        "files": "encrypt-actions/*",
         "rules": {
           "node/no-unsupported-features/node-builtins": "off"
         }
